feat: interval screen -> interval active로 이동하는 flow ios, watchos 통일
