+++ Projects/Presentation/Sources/Flows/Interval/Screen/IntervalScreen+iOS.swift
@@ -25,7 +25,7 @@ public extension IntervalScreen {
                     viewModel.nextScreen()
                 }
                 .sheet(isPresented: $viewModel.isBottomSheetPresent) {
-                    addIntervalScreen
+                    AddIntervalScreen(viewModel: intervalDIContainer.addIntervalDependencies(intervalRouter: router))
                 }
         }
     }
