+++ Projects/Presentation/Sources/Flows/Interval/Screen/IntervalScreen.swift
@@ -12,19 +12,19 @@ public struct IntervalScreen: View {
     @StateObject var router: IntervalRouter
     @StateObject var viewModel: IntervalViewModel
     
+    let intervalDIContainer: IntervalDIContainerInterface
     let intervalListScreen: IntervalListScreen
-    let addIntervalScreen: AddIntervalScreen
     
     public init(
         intervalDIContainer: IntervalDIContainerInterface
     ) {
+        self.intervalDIContainer = intervalDIContainer
         let router = intervalDIContainer.intervalRouter()
         
         self._router = .init(wrappedValue: router)
         self._viewModel = .init(wrappedValue: intervalDIContainer.intervalScreenDependencies(intervalRouter: router))
         
         self.intervalListScreen = .init(viewModel: intervalDIContainer.intervalListDependencies(intervalRouter: router))
-        self.addIntervalScreen = .init(viewModel: intervalDIContainer.addIntervalDependencies(intervalRouter: router))
     }
 
     public var body: some View {
