--- Projects/Presentation/Sources/Interval/Active/IntervalInfoActiveView.swift
@@ -1,30 +1,0 @@
-////
-////  WatchIntervalInfoView.swift
-////  WatchPresentation
-////
-////  Created by 최유경 on 11/18/23.
-////
-//
-//import SwiftUI
-//
-//import SharedDesignSystem
-//
-//struct IntervalInfoActiveView: View {
-//    var body: some View {
-//        HStack{
-//            VStack(alignment: .leading){
-//                Text("인터벌")
-//                Text("심박수")
-//                Text("칼로리")
-//                Text("진행시간")
-//            }
-//            .foregroundColor(Color.textColor50)
-//            .padding(.horizontal,12)
-//            Spacer()
-//        }
-//    }
-//}
-//
-//#Preview {
-//    IntervalInfoActiveView()
-//}
