+++ Projects/Presentation/Sources/Interval/List/IntervalListScreen+watchOS.swift
@@ -17,19 +17,10 @@ public extension IntervalListScreen {
                     //TODO: 셀로 변경 해야함
                     Text(item.title)
                 }
-                
-                Button(action: {
-                    viewModel.tapIntervalDetailPageButton()
-                }, label: {
-                    Text("인터벌 상세페이지 가기")
-                })
             }
             .padding(.horizontal, 24)
             .padding(.top, 32)
         }
-        .sheet(isPresented: $viewModel.showEditIntervalView, content: {
-            EditIntervalScreen()
-        })
         .onAppear() {
             viewModel.fetchIntervalItems()
         }
